create interface for product repository
