+++ src/main/java/com/security/jdbc/product/service/ProductServiceImpl.java
@@ -2,15 +2,16 @@ package com.security.jdbc.product.service;
 
 
 import com.security.jdbc.product.dto.request.CreateProductDtoRequest;
+import com.security.jdbc.product.repository.IProductRepository;
 import com.security.jdbc.product.repository.ProductRepository;
 import org.springframework.stereotype.Service;
 
 @Service
-public class ProductService {
+public class ProductService implements IProductService {
 
-    private final ProductRepository productRepository;
+    private final IProductRepository productRepository;
 
-    public ProductService(ProductRepository productRepository){
+    public ProductService(IProductRepository productRepository){
         this.productRepository = productRepository;
     }
